requirements.txt
PyQt6==6.6.1
pandas==2.1.4
openpyxl==3.1.2
PyYAML==6.0.1
xlrd==2.0.1
pyarrow==14.0.2

